requirements.txt
streamlit
tensorflow
numpy
matplotlib
pillow
streamlit-drawable-canvas